cx_autotest.py: update for new repo
The old monorepo edited autotest.py and vehicleinfo.py directly and
placed an extra module (carbonix.py) in the same directory with it. This
is not in-line with the new repo structure. Unfortunately, the way
autotest.py is written, there's no good way to use it as an externally
imported module, so we need to duplicate some of its functionality here.
The advantage is that now the autotest is one file now.

Additionally, I picked the timeout for the CHT to be just barely long
enough and it was sometimes failing, falling <1C below the threshold.