features.inc: sort entries
This sorts the defines alphabetically (kinda) in the same order that
extract_features.py does. It replaces "HAL_" with "AP_" for sorting
purposes only, so that related AP_ and HAL_ entries stay together.

This is a non-functional change. You can see this easily by purely
sorting the undef and define sections with VSCode and diffing.